+++ derive_builder/src/options/struct_options.rs
@@ -7,7 +7,6 @@ use options::DefaultExpression;
 pub struct StructOptions {
     /// Whether or not this struct should implement its own build method.
     pub build_fn_enabled: bool,
-    
     /// The name of the emitted build method.
     pub build_fn_name: syn::Ident,
     /// Name of the builder struct, e.g. `FooBuilder`.
@@ -33,6 +32,9 @@ pub struct StructOptions {
     pub bindings: Bindings,
     /// Default expression for the whole struct, e.g. `#[builder(default)]` (default to None).
     pub default_expression: Option<DefaultExpression>,
+    /// Path to the optional validation function to invoke before the
+    /// macro-generated `build` method executes.
+    pub validate_fn: Option<syn::Path>,
 }
 
 impl StructOptions {
@@ -66,6 +68,7 @@ impl StructOptions {
             default_struct: self.default_expression
                 .as_ref()
                 .map(|x| { x.parse_block(self.bindings.no_std) }),
+            validate_fn: self.validate_fn.as_ref(),
         }
     }
 }